move types to matching filename
